+++ docs/api/changelog.rst
@@ -110,6 +110,9 @@ Version 0.9
 * Fix :attr:`aioxmpp.RosterClient.groups` not being updated when items are
   removed during initial roster update.
 
+* :meth:`aioxmpp.RosterClient.on_group_added`,
+  :meth:`~aioxmpp.RosterClient.on_group_removed`.
+
 .. _api-changelog-0.8:
 
 Version 0.8